Add Menu controller, service, to

+++ src/main/java/ru/erulaev/restaurantvoting/app/AuthUser.java
@@ -15,7 +15,7 @@ public class AuthUser extends org.springframework.security.core.userdetails.User
         this.user = user;
     }
 
-    public int id() {
+    public long id() {
         return user.id();
     }
