+++ src/main/java/ru/erulaev/restaurantvoting/common/HasId.java
@@ -5,16 +5,16 @@ import org.springframework.util.Assert;
 
 public interface HasId {
 
-    Integer getId();
+    Long getId();
 
-    void setId(Integer id);
+    void setId(Long id);
 
     @Schema(hidden = true)
     default boolean isNew() {
         return getId() == null;
     }
 
-    default int id() {
+    default long id() {
         Assert.notNull(getId(), "Entity must has id");
         return getId();
     }
