+++ src/main/java/ru/erulaev/restaurantvoting/common/model/BaseEntity.java
@@ -18,9 +18,9 @@ public abstract class BaseEntity implements HasId {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Schema(accessMode = Schema.AccessMode.READ_ONLY) // https://stackoverflow.com/a/28025008/548473
-    protected Integer id;
+    protected Long id;
 
-    //    https://stackoverflow.com/questions/1638723
+    // https://stackoverflow.com/questions/1638723
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
