+++ src/main/java/ru/erulaev/restaurantvoting/common/to/NamedTo.java
@@ -15,7 +15,7 @@ public class NamedTo extends BaseTo {
     @NoHtml
     protected String name;
 
-    public NamedTo(Integer id, String name) {
+    public NamedTo(Long id, String name) {
         super(id);
         this.name = name;
     }
