+++ src/main/java/ru/erulaev/restaurantvoting/user/repository/VoteRepository.java
@@ -5,6 +5,6 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.erulaev.restaurantvoting.user.model.Vote;
 
 @Transactional(readOnly = true)
-public interface VoteRepository extends JpaRepository<Vote, Integer> {
+public interface VoteRepository extends JpaRepository<Vote, Long> {
 
 }
